refactor: 댓글 기능 분리
