+++ src/entities/comment/api/fetchComments.ts
@@ -1,0 +1,12 @@
+  // 댓글 가져오기
+  const fetchComments = async (postId) => {
+    if (comments[postId]) return // 이미 불러온 댓글이 있으면 다시 불러오지 않음
+    try {
+      const response = await fetch(`/api/comments/post/${postId}`)
+      const data = await response.json()
+      setComments(postId, data.comments)
+    } catch (error) {
+      console.error("댓글 가져오기 오류:", error)
+    }
+  }
+
